+++ netbox_dhcp_kea_plugin/template_content.py
@@ -1,7 +1,7 @@
 from ipam.models import Prefix
 from netbox.plugins import PluginTemplateExtension
 
-from .models import PrefixDHCPConfig
+from .models import Subnet
 
 
 class PrefixDHCPInfo(PluginTemplateExtension):
@@ -23,7 +23,7 @@ class PrefixDHCPInfo(PluginTemplateExtension):
 
         try:
             dhcp_config = (
-                PrefixDHCPConfig.objects.select_related("server")
+                Subnet.objects.select_related("server")
                 .prefetch_related("option_data", "client_classes")
                 .get(prefix=obj)
             )
@@ -34,7 +34,7 @@ class PrefixDHCPInfo(PluginTemplateExtension):
                     "dhcp_config": dhcp_config,
                 },
             )
-        except PrefixDHCPConfig.DoesNotExist:
+        except Subnet.DoesNotExist:
             return ""